feat(project): Ensure VS Code IDE configured consistently
Often devs have their IDE configured differently, we should mandate
the usage of the same IDE also for consistency as this can play
havoc on code formatting etc. To this end I have added config for
the VS Code IDE, to recommend mandatory extensions, if they are not
already installed and to ensure that code is linted and formatted
when the file is saved.

IssueID EODHP-560