fix(project): Update ESLint to use shared config
I've updated the ESLint rules to be the same as what the resource
catalogue uses, where applicable. This reduces the amount of config
needed and makes the code a bit more consistent.

I also added prettier which will format the code to a standard format
and this is enacted whenever you save the page in VS Code, so all
JavaScript is consistently laid out.

IssueID EODHO-560