fix(frontend): Fix Lint issues and reformat code
I have fixed any linting issues found and reformatted the code so
it is consistent.

I had to rename the files to *.mjs, this is ES6/Module parlance instead
of the CommonJS that webpack uses by default. Basically in a CommonJS
you cannot use the `import`, which is heavily used. NodeJS is moving
away from CommonJS, but not everything has caught up yet, these things
take time.

IssueID EODHP-560