fix(frontend): Rename JavaScript files
I initially had trouble using `import` as this is an ES Module tool
whereas the code was done in `CommonJS`. This fixed some linting
issues I had, but subsequent changes seem to have made that unnecessary
and having webpack named like this caused issues when building the
docker image. Webpack hasn't quite caught up with the ES Module
setup yet, so renaming the files seems to have fixed this.