fix(project): Update reference to eslint file
I'm not exactly sure why we are copying the linting code to Docker
as it is an aid to developers while writing code, not when it is
built or running. However, I didn't want to make too many changes
either. I had change the linting to use JavaScript config file rather
than JSON and had forgotten to update the reference to it here.