fix(styles): Undo moving of import
Importing files other than at the top of a file is frowned upon, however
there are situations where it is necessary, like this one, where we
are overriding some of `bootstraps` SASS Variables. After discussion
in PR, we agreed to revert this change and ignore the linting rule
as the rule is generally the right thing.