feat(project): Add ability to run docker locally with live-reload
It is quite annoying not being able to see changes within the context
of the `web presence` and the `resource catalogue` locally, prior
to deploying to `dev`. I have now added Docker related files so we
can run it locally, so as to see the app in it's full context.