fix(sign_in): Improve code quality
During review it was suggested to make the code more DRY, by seting
the `redirect_tag` to the referer or an empty string, as opposed to
the if/else I had, this is much neater.

It was also raised that getting the referer from the headers dictionarty
can be improved as well. I had tried before but didn't realise the keys
used in `request.META` and in `request.headers` differ, I had been
looking for `HTTP_REFERER`, but it is simply `Referer`.