Made styling inline.
I tried setting the styles in the SCSS files by targeting the id, however this did not seem to work as expected as the style was not affecting the element. I moved the styling to inline as this was the only way that I could get the height to behave as I wanted. I was changing the css in the welcome_page.css and also tried the home.scss file and running python mange.py collectstatic though none of this fixed the issue.