fix(assets): Prevent importing styles multiple times
Looking at the `<head />` section, I noticed the styles were being
imported multiple times, which while it didn't break anything, made
it harder to turn styles on/off when testing something in the browser.
It seems to be caused by the `entrypoint.js` script importing multiple
SCSS files, so I wrapped them all in a single file and import only
that one and it appears to have fixed the issue.

IssueID EODHP-596