fix(project): Prevent prettier formatting HTML
Using Django templated HTML meant that prettier could re-format the
file breaking the templating by putting all the pragmas on a single
line. If a template was in say `templates/home/index.html`, then
prettier still formatted it as the line in the `.prettierignore` file
was too specific, only files inside `templates` directories where
ignored, not those in sub-directories.